Import gcc-3.4.6

+++ gcc/testsuite/gcc.dg/i386-sse-5.c
@@ -1,6 +1,8 @@
-/* { dg-do compile { target i?86-*-* } } */
+/* { dg-do compile { target i?86-*-* x86_64-*-* } } */
+/* { dg-require-effective-target ilp32 } */
 /* { dg-options "-Winline -O2 -march=i386" } */
-typedef int v2df __attribute__ ((mode(V2DF)));
+
+typedef float v2df __attribute__ ((mode(V2DF)));
 v2df p;
 q(v2df t)
 {			 /* { dg-warning "SSE" "" } */
